backend/requirements.txt
fastapi
fastapi[standard]
uvicorn
pydentic
pydantic[email]
sqlalchemy
cryptography
passlib[argon2]
argon2-cffi
email-validator
PyJWT